Changed translation logic.
When looking at a page and choosing the other language, the blog tried
to translate the page. There was a problem related to page names (which
are also translated) and how to identify them.

That functionality was removed and when clicking a language it just goes
to the homepage in the selected language. This simplifies a great deal
the logic and management of the blog which is more important for me.